prosopopee: add ratio property for videos
To prepare for an alignment fix that requires to know the ratio of the
video screenshot, let's add a ratio property.

This property is calling `ffprobe` or `avprobe` depending on whether
ffmpeg or libav-tools was installed. It is not adding a new dependency.

The output of this particular `ffprobe` or `avprobe` returns the width
and height of the video, separated by a comma.

Signed-off-by: Quentin Schulz <foss@0leil.net>