themes: pictures-group: fix alignment when pictures do not fill the flex container
In most cases, a line in a pictures-group contains enough pictures or at
least in the correct format (usually in landscape mode) to actually make
the browser call flex-shrink.

However, in a few specific cases, when pictures do not fill the flex
container, there is some space left on the right side of the line which
could be filled in.

In that case, we can use the flex-grow property. To be able to keep
everything aligned, we need to set the flex-grow property according to
the ratio of the picture so that all pictures in the line are grown
proportionally.

Adding this alignment fix required a bit of re-ordering so we could call
the ratio property for the div of class `picture`.

Fixes #92

Signed-off-by: Quentin Schulz <foss@0leil.net>