Fix unassigned variable error
If exif data exists but 'DateTime' is not in exif_data, the datetime
that we're supposed to return never gets assigned and triggers an Error.

We first attempt to populate the exif_data and then IF it's populated we
use it ELSE we use getmtime so it's always one or the other. No error
here because exif_data is already initialized as {}.

resolves #112