prosopopee: disable og if url is not set in root settings.yaml
OpenGraph only works if it's provided the URL of the website.

Moreover, if the root settings.yaml is missing url setting but has og
enabled, the build will fail.

Let's fix that by disabling og if url is not set in root settings.yaml
and print a warning when this happens.

Signed-off-by: Quentin Schulz <foss@0leil.net>