prosopopee: cache: compare cached options and current options after curation
json.dumps() which is used to write the cache dict to a file transforms
tuples into a list. With the current implementation, if a tuple is
supposed to be cached, the needs_to_be_generated method will always
return True even though it might not be correct.

In order to support tuples in cache entries, let's pass the options
passed as parameter to the method through json.loads(json.dumps()) to
have the same format between cached options and to-be-compared options.

This will be used in a later commit which adds a tuple (width, height) to
the cache.

Signed-off-by: Quentin Schulz <foss@0leil.net>