prosopopee: themes: exposure: full-picture: fix deformed thumbnails for videos
full-picture class used to fit the whole screen which is fine for photos
since they're cropped.

However when a video is to be put in full-picture class, its width is
limited to the one of the display though its thumbnail still has its
width and height set to 100%.

This makes the thumbnail not respect its original ratio and results in
in the video being surrounded by deformed parts of the thumbnail (the
thumbnail taking 100% of the display height while the video only takes
whatever is needed to respect the ratio with the width being 100% of
the display's).

Fixes #138.

Signed-off-by: Quentin Schulz <foss@0leil.net>