themes: exposure: style.css: fill space when video used as cover in gallery index pages
When a video is used as a cover of a gallery, the current behavior is
the video to be limited in width or height of the parent div. However
this means cover videos (portrait ones particularly) won't fill the
parent div, as do cover images.

Let's make the style consistent by having the cover video (and its
thumbnail) fill the parent div.

Signed-off-by: Quentin Schulz <foss@0leil.net>